+++ core/realtime-core/src/main/kotlin/org/http4k/sse/SseEventId.kt
@@ -1,0 +1,4 @@
+package org.http4k.sse
+
+@JvmInline
+value class SseEventId(val value: String)