+++ core/realtime-core/src/test/kotlin/org/http4k/sse/SseMessageTest.kt
@@ -22,7 +22,7 @@ class SseMessageTest {
     @Test
     fun `converts event toMessage`() {
         assertRoundtrip(
-            SseMessage.Event("event", "data1\ndata2", "id"), """event: event
+            SseMessage.Event("event", "data1\ndata2", SseEventId("id")), """event: event
 data: data1
 data: data2
 id: id