+++ pro/mcp/sdk/src/main/kotlin/org/http4k/mcp/protocol/SessionId.kt
@@ -1,9 +1,11 @@
 package org.http4k.mcp.protocol
 
+import dev.forkhandles.values.NonBlankStringValueFactory
+import dev.forkhandles.values.StringValue
 import dev.forkhandles.values.UUIDValue
 import dev.forkhandles.values.UUIDValueFactory
 import java.util.UUID
 
-class SessionId private constructor(value: UUID) : UUIDValue(value) {
-    companion object : UUIDValueFactory<SessionId>(::SessionId)
+class SessionId private constructor(value: String) : StringValue(value) {
+    companion object : NonBlankStringValueFactory<SessionId>(::SessionId)
 }