+++ pro/mcp/sdk/src/main/kotlin/org/http4k/mcp/server/ws/McpWsHandler.kt
@@ -6,20 +6,24 @@ import org.http4k.core.Request
 import org.http4k.format.jsonRpcRequest
 import org.http4k.mcp.server.RealtimeMcpProtocol
 import org.http4k.mcp.util.McpJson
+import org.http4k.routing.bindWs
 import org.http4k.routing.websockets
 import org.http4k.sse.SseMessage
 import org.http4k.websocket.WsMessage
+import org.http4k.websocket.WsResponse
 
 /**
  * This Websocket handler can be bound to whatever path is required by the server with
  * ws("/path" bind <McpWsHandler>
  */
-fun McpWsHandler(mcpProtocol: RealtimeMcpProtocol) = websockets {
-    val newSessionId = mcpProtocol.newSession(it)
-    it.onMessage {
-        val req = Request(POST, "/ws").body(it.bodyString())
-        mcpProtocol(newSessionId, Body.jsonRpcRequest(McpJson).toLens()(req), req)
-    }
+fun McpWsHandler(mcpProtocol: RealtimeMcpProtocol) = websockets("" bindWs { req ->
+    WsResponse {
+        val newSessionId = mcpProtocol.newSession(req, it)
+        it.onMessage {
+            val newReq = Request(POST, "/ws").body(it.bodyString())
+            mcpProtocol(newSessionId, Body.jsonRpcRequest(McpJson).toLens()(newReq), req)
+        }
 
-    it.send(WsMessage(SseMessage.Event("endpoint", "").toMessage()))
-}
+        it.send(WsMessage(SseMessage.Event("endpoint", "").toMessage()))
+    }
+})