+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 This list is not intended to be all-encompassing - it will document major and breaking API
 changes with their rationale when appropriate:
 
-### uncut
+### v6.1.0.0 (uncut)
+- **http4k-*** : Upgrade some dependency versions.
+- **http4k-realtime-core** :[Breaking] SseEventId is now a type instead of just a string
 - **http4k-connect-amazon-ses***: [Breaking] Upgrade to SES V2 API. H/T @oharaandrew314
 
 ### v6.0.1.0